refactor: rename set_time module to time
